pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-browser"
version = "0.1.0"
description = "A local-first agentic browser that renders synthesized webpages."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116,<1.0",
  "uvicorn>=0.35,<1.0",
  "pydantic-settings>=2.10,<3.0",
  "jinja2>=3.1,<4.0",
  "httpx>=0.28,<1.0",
  "beautifulsoup4>=4.13,<5.0",
  "python-dotenv>=1.1,<2.0",
  "langgraph>=0.2,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
